# app/static/openapi.yaml
openapi: "3.0.2"
info:
  title: Vending Machine
  version: "6.9"
servers:
  - url: /
...
paths:
...
  /products/:
    description: 'get all products'
    get:
      tags:
        - Products
      responses:
        200:
          description: 'product table are not empty'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sample_products'
        204:
          description: 'table is empty'
        400:
          description: 'bad request, something was wrong with query strings'
        405:
          description: 'method not allowed'
  /add_products/:
    description: 'get all products'
    parameters:
        - in: query
          name: product_name
          schema:
            type: string
          description: The name of the product
        - in: query
          name: product_code
          schema:
            type: integer
          description: The code of the product use to uniquely idendify each of them
        - in: query
          name: product_quantity
          schema:
            type: integer
          description: The quantity of the product
        - in: query
          name: price_per_unit
          schema:
            type: integer
          description: The price of the product
    post:
      tags:
        - Products
      responses:
        200:
          description: 'successfully added'
        204:
          description: 'table is empty'
        400:
          description: 'bad request, something was wrong with query strings'
        405:
          description: 'method not allowed'
  /edit_products/:
    description: 'get all vending machines'
    parameters:
        - in: query
          name: id
          schema:
            type: integer
          description: The id of the product
        - in: query
          name: product_name
          schema:
            type: string
          description: The name of the product
        - in: query
          name: product_code
          schema:
            type: integer
          description: The code of the product use to uniquely idendify each of them
        - in: query
          name: product_quantity
          schema:
            type: integer
          description: The quantity of the product
        - in: query
          name: price_per_unit
          schema:
            type: integer
          description: The price of the product
    post:
      tags:
        - Products
      responses:
        200:
          description: 'successfully added'
        204:
          description: 'table is empty'
        400:
          description: 'bad request, something was wrong with query strings'
        405:
          description: 'method not allowed'
  /delete_products/:
    description: 'get all vending machines'
    parameters:
        - in: query
          name: id
          schema:
            type: integer
          description: The id of the product
    delete:
      tags:
        - Products
      responses:
        200:
          description: 'successfully added'
        204:
          description: 'table is empty'
        400:
          description: 'bad request, something was wrong with query strings'
        405:
          description: 'method not allowed'
...
components:
  schemas:
    sample_products:
      type: object
      example: [{
        "id": 1,
        "price_per_unit": 1.0,
        "product_code": 0,
        "product_name": "test_product_1",
        "product_quantity": 1
        },
        {
        "id": 2,
        "price_per_unit": 1.0,
        "product_code": 1,
        "product_name": "test_product_2",
        "product_quantity": 1
        }
        ]
...